content/home: fold DIY/DIY+ into NERD and rename PRO to KAFI
the DIY/DIY+ concepts were conflicting with what they represented when
compared to PRO. some said it should've actually been the opposite: PRO
is the topmost for DIY kind of stuff... you see the ambiguitiy.

so, we came up with new names and simplified: NERD is for people who'd
want to mess around on their own and KAFI is the plug-and-play coffee maker.